requirements.txt
langchain
pypdf
unstructured
sentence_transformers
pinecone-client
llama-cpp-python
numpy==1.23.4
huggingface_hub
beautifulsoup4
python-dotenv
cohere
pandas
seaborn
matplotlib
wikipedia
neo4j
py2neo